+++ AccelByte.Sdk/Api/Eventlog/Operation/EventV2/GetPublicEditHistory.cs
@@ -57,9 +57,9 @@ namespace AccelByte.Sdk.Api.Eventlog.Operation
             
             public string? EndDate { get; set; }
             
-            public double? Offset { get; set; }
+            public long? Offset { get; set; }
             
-            public double? PageSize { get; set; }
+            public long? PageSize { get; set; }
             
             public string? StartDate { get; set; }
             
@@ -74,13 +74,13 @@ namespace AccelByte.Sdk.Api.Eventlog.Operation
                 return this;
             }
 
-            public GetPublicEditHistoryBuilder SetOffset(double _offset)
+            public GetPublicEditHistoryBuilder SetOffset(long _offset)
             {
                 Offset = _offset;
                 return this;
             }
 
-            public GetPublicEditHistoryBuilder SetPageSize(double _pageSize)
+            public GetPublicEditHistoryBuilder SetPageSize(long _pageSize)
             {
                 PageSize = _pageSize;
                 return this;
@@ -137,8 +137,8 @@ namespace AccelByte.Sdk.Api.Eventlog.Operation
             string namespace_,            
             string userId,            
             string? endDate,            
-            double? offset,            
-            double? pageSize,            
+            long? offset,            
+            long? pageSize,            
             string? startDate,            
             string? type            
         )
