+++ AccelByte.Sdk/Api/Eventlog/Operation/EventV2/GetUserEventsV2Public.cs
@@ -31,9 +31,9 @@ namespace AccelByte.Sdk.Api.Eventlog.Operation
             
             public string? EventName { get; set; }
             
-            public double? Offset { get; set; }
+            public long? Offset { get; set; }
             
-            public double? PageSize { get; set; }
+            public long? PageSize { get; set; }
             
             public string? StartDate { get; set; }
             
@@ -52,13 +52,13 @@ namespace AccelByte.Sdk.Api.Eventlog.Operation
                 return this;
             }
 
-            public GetUserEventsV2PublicBuilder SetOffset(double _offset)
+            public GetUserEventsV2PublicBuilder SetOffset(long _offset)
             {
                 Offset = _offset;
                 return this;
             }
 
-            public GetUserEventsV2PublicBuilder SetPageSize(double _pageSize)
+            public GetUserEventsV2PublicBuilder SetPageSize(long _pageSize)
             {
                 PageSize = _pageSize;
                 return this;
@@ -110,8 +110,8 @@ namespace AccelByte.Sdk.Api.Eventlog.Operation
             string userId,            
             string? endDate,            
             string? eventName,            
-            double? offset,            
-            double? pageSize,            
+            long? offset,            
+            long? pageSize,            
             string? startDate            
         )
         {
